Updated the uninstall script.

+++ uninstall.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
-while read p; do
-    rm $p
-done < install_manifest.txt
+installPath="install_manifest.txt"
+
+if [ -f "$installPath" ]; then
+    while read p; do
+        rm $p
+    done < $installPath
+    rm "$installPath"
+else
+    echo "Cannot uninstall: $installPath does not exist."
+fi
